requirements.txt
agate==1.7.1
aiobotocore==2.13.2
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aioitertools==0.11.0
aiosignal==1.3.1
airbyte @ file:///E:/R%26D/flask-hubspot-app/hubspot-app/v3Hubspot/hubspotv3/PyAirbyte
airbyte-api==0.49.4
airbyte-cdk==4.6.2
airbyte_protocol_models_dataclasses==0.13.0
airbyte_protocol_models_pdv2==0.12.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attributes-doc==0.4.0
attrs==23.2.0
awscli==1.34.0
Babel==2.15.0
backoff==2.2.1
BareNecessities==0.2.8
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.111
botocore==1.35.0
bracex==2.5
browsermob-proxy==0.8.0
bs4==0.0.2
cachetools==5.3.3
cattrs==24.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedrivermanager==0.0.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
connectedpapers-py==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dacite==1.8.1
daff==1.3.46
dataclasses-json==0.6.7
DateTime==5.4
db-dtypes==1.2.0
dbt==1.0.0.38.6
dbt-adapters==1.2.1
dbt-bigquery==1.8.1
dbt-common==1.1.0
dbt-core==1.7.17
dbt-extractor==0.5.1
dbt-mysql==1.7.0
dbt-postgres==1.8.0
dbt-redshift==1.8.0
dbt-semantic-interfaces==0.4.4
dbt-snowflake==1.8.2
dbt-trino==1.8.0
debugpy==1.8.1
decorator==5.1.1
Delorean==1.0.0
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.3
docutils==0.16
dpath==2.2.0
duckdb==1.0.0
duckdb_engine==0.13.2
executing==2.0.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.12.2
genson==1.2.2
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.23.1
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.2
hubspot==0.1.14.dev0
hubspot-api-client==9.0.0
humanize==4.10.0
idna==3.6
image==1.5.33
ImageHash==4.3.1
importlib-metadata==6.11.0
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.1.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==0.2
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==24.3.1
kiwisolver==1.4.5
langchain-core==0.1.42
langsmith==0.1.116
leather==0.4.0
Logbook==1.5.3
lxml==5.1.0
Mail==2.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mashumaro==3.13
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
missingno==0.5.2
mmhash3==3.0.1
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
multimethod==1.12
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandas-profiling==3.2.0
parse==1.20.1
parsedatetime==2.6
parso==0.8.4
pathspec==0.11.2
pdfminer.six==20231228
pdoc==14.6.1
pendulum==2.1.2
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.17.2
pyiceberg==0.6.1
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL==24.1.0
pyparsing==3.1.4
pypdfium2==4.30.0
pypiwin32==223
pyrate-limiter==3.1.1
pyrsistent==0.20.0
PySocks==1.7.1
pyspark==3.5.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-ulid==2.7.0
pytimeparse==1.1.8
pytz==2024.1
pytzdata==2020.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.2
redshift-connector==2.0.917
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-cache==1.2.1
rich==13.8.0
rpds-py==0.18.1
rsa==4.7.2
s3fs==2024.6.1
s3transfer==0.10.1
scramp==1.4.5
selenium==4.20.0
serpyco-rs==1.11.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.10.1
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.51
sqlalchemy-bigquery==1.9.0
sqlparse==0.5.0
stack-data==0.6.3
strictyaml==1.7.3
structlog==24.4.0
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
text-unidecode==1.3
tomlkit==0.12.5
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
trino==0.328.0
trio==0.25.0
trio-websocket==0.11.1
types-PyYAML==6.0.12.20240808
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
url-normalize==1.4.3
urllib3==1.26.19
uuid7==0.1.0
virtualenv==20.25.1
visions==0.7.4
wcmatch==8.4
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.2
zope.interface==6.2
